configmap_source: add KubernetesConfigMap source
We can use this to load a Kubernetes ConfigMap mounted volume as
a conf.Source.

I hope to expand on this to allow people to subscribe to values so
that running programs can get notified about updated configuration
values. However this is less of a priority as the first application is
to use them from Terraform, which does not change that much.

Also make it possible to edit the default Loader - without these
it's difficult to get, e.g. "the default behavior plus a Kubernetes
volume".

Add a Subscriber interface that can be used to watch a set of config
values and get notified of changes to them.